refactor: Split DTXAPIClientTests and fix MetronomeTimingEngine issues
- Split 359-line DTXAPIClientTests into focused smaller suites to satisfy SwiftLint type_body_length limit:
  * DTXAPIClientTests: Core functionality tests (66 lines)
  * DTXAPIClientInitTests: Initialization and configuration tests
  * DTXAPIClientURLTests: URL handling and validation tests
  * DTXAPIClientConcurrencyTests: Async and concurrency tests
- Remove unused logger property from MetronomeTimingEngine
- Fix startAtTime() test-mode branch to respect provided startTime parameter
- Update simulateTestBeat() to accept startTime parameter for proper timing API support

Maintains test coverage while improving code organization and satisfying linting requirements.
