+++ VirgoTests/DTXAPIClientTests.swift
@@ -9,7 +9,9 @@ import Testing
 import Foundation
 @testable import Virgo
 
-@Suite("DTX API Client Tests", .serialized)
+/// Core DTX API Client functionality tests
+/// Additional tests split into: DTXAPIClientInitTests, DTXAPIClientURLTests, DTXAPIClientConcurrencyTests
+@Suite("DTX API Client Core Tests", .serialized)
 struct DTXAPIClientTests {
     
     init() async throws {
@@ -57,300 +59,6 @@ struct DTXAPIClientTests {
         try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
         let resetURL = client.baseURL
         #expect(resetURL == "http://127.0.0.1:8001")
-        
-        // Clean up after test
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-    }
-    
-    @Test("DTXAPIClient handles custom server URLs")
-    func testCustomServerURLs() async throws {
-        let client = DTXAPIClient()
-        
-        // Clean up any existing value
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        
-        let testURLs = [
-            "http://localhost:3000",
-            "https://api.example.com",
-            "http://192.168.1.100:8001",
-            "https://dtx-server.herokuapp.com"
-        ]
-        
-        for url in testURLs {
-            client.setServerURL(url)
-            UserDefaults.standard.synchronize()
-            try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-            #expect(client.baseURL == url)
-        }
-        
-        // Clean up after test
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-    }
-    
-    @Test("DTXAPIClient test connection handles invalid URLs gracefully")
-    func testConnectionWithInvalidURL() async throws {
-        // Ultimate Infrastructure Transcendence: Root cause resolution for 0.000 second failures
-        await UltimateInfrastructureTranscendence.shared.transcendInfrastructureBarriers(for: "testConnectionWithInvalidURL")
-        
-        // Precision-Calibrated: Adaptive network isolation for True 100% success rate
-        let environment = await AdaptiveTestIsolation.shared.createAdaptiveEnvironment(for: "testConnectionWithInvalidURL")
-        await PrecisionHardwareMitigation.shared.applyPrecisionMitigation(for: "testConnectionWithInvalidURL")
-        
-        let optimalDelay = await TestExecutionManager.shared.getOptimalDelay(for: "testConnectionWithInvalidURL")
-        try await Task.sleep(nanoseconds: UInt64(optimalDelay * 1_000_000_000))
-        
-        let client = DTXAPIClient()
-        
-        // Revolutionary optimal UserDefaults cleanup
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 250_000_000) // 250ms - optimal delay
-        
-        client.setServerURL("invalid-url-format")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 50_000_000) // 50ms - optimal delay
-        
-        // Ultra-advanced TaskGroup with precision timing for 99%+ breakthrough
-        let connectionResult = await withTaskGroup(of: Bool.self) { group in
-            group.addTask {
-                await client.testConnection()
-            }
-            
-            // Ultra-precision timeout protection with advanced cancellation
-            group.addTask {
-                try? await Task.sleep(nanoseconds: 8_000_000_000) // 8 seconds - ultra-precision
-                return false // Timeout fallback
-            }
-            
-            // Ultra-advanced result handling with explicit cancellation
-            let result = await group.next() ?? false
-            group.cancelAll() // Ensure clean cancellation
-            return result
-        }
-        #expect(connectionResult == false)
-        
-        // Precision-calibrated cleanup with adaptive network stabilization
-        await PrecisionHardwareMitigation.shared.applyPrecisionMitigation(for: "testConnectionWithInvalidURL")
-        await AdaptiveTestIsolation.shared.performAdaptiveCleanup(for: environment)
-        
-        // Clean up after test with hardware-aware timing
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-    }
-    
-    @Test("DTXAPIClient error descriptions are meaningful")
-    func testErrorDescriptions() {
-        let invalidURLError = DTXAPIError.invalidURL
-        let noDataError = DTXAPIError.noData
-        let decodingError = DTXAPIError.decodingError
-        let networkError = DTXAPIError.networkError(URLError(.notConnectedToInternet))
-        
-        #expect(invalidURLError.errorDescription == "Invalid server URL")
-        #expect(noDataError.errorDescription == "No data received from server")
-        #expect(decodingError.errorDescription == "Failed to decode server response")
-        #expect(networkError.errorDescription?.contains("Network error") == true)
-    }
-    
-    @Test("DTXAPIClient URL construction for various endpoints")
-    func testURLConstruction() async throws {
-        let client = DTXAPIClient()
-        
-        // Clean up any existing value
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        
-        client.setServerURL("http://test-server.com:8080")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        
-        // Verify the URL was actually set in UserDefaults
-        let storedURL = UserDefaults.standard.string(forKey: "DTXServerURL")
-        #expect(storedURL == "http://test-server.com:8080", "URL should be stored in UserDefaults")
-        
-        // Test internal URL construction logic (indirectly)
-        let baseURL = client.baseURL
-        #expect(baseURL == "http://test-server.com:8080")
-        
-        // Would construct URLs like:
-        // "\(baseURL)/dtx/list" -> "http://test-server.com:8080/dtx/list"
-        // "\(baseURL)/dtx/download/filename.dtx" -> "http://test-server.com:8080/dtx/download/filename.dtx"
-        // etc.
-        
-        // Clean up after test
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-    }
-    
-    @Test("DTXAPIClient handles empty and nil server URLs")
-    func testEmptyServerURLs() async throws {
-        let client = DTXAPIClient()
-        
-        // Clean up any existing value
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        
-        // Empty string should fall back to default
-        client.setServerURL("")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        #expect(client.baseURL == "http://127.0.0.1:8001")
-        
-        // Reset should also fall back to default
-        client.setServerURL("http://custom.com")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        client.resetToLocalServer()
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        #expect(client.baseURL == "http://127.0.0.1:8001")
-        
-        // Clean up after test
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-    }
-    
-    @Test("DTXAPIClient session configuration is correct")
-    func testSessionConfiguration() {
-        let client = DTXAPIClient()
-        let session = client.session
-        
-        #expect(session.configuration.timeoutIntervalForRequest == 30.0)
-        #expect(session.configuration.timeoutIntervalForResource == 60.0)
-        #expect(session.configuration.httpMaximumConnectionsPerHost == 2)
-        #expect(session.configuration.requestCachePolicy == .reloadIgnoringLocalCacheData)
-        #expect(session.configuration.waitsForConnectivity == true)
-        #expect(session.configuration.allowsConstrainedNetworkAccess == true)
-        #expect(session.configuration.allowsExpensiveNetworkAccess == true)
-    }
-    
-    @Test("DTXAPIClient state management") 
-    func testStateManagement() async {
-        let client = DTXAPIClient()
-        
-        // Initial state
-        #expect(client.isLoading == false)
-        #expect(client.errorMessage == nil)
-        
-        // Note: We can't easily test the private updateLoadingState method,
-        // but we can verify the Published properties exist and are accessible
-        await MainActor.run {
-            client.isLoading = true
-            client.errorMessage = "Test error"
-        }
-        
-        #expect(client.isLoading == true)
-        #expect(client.errorMessage == "Test error")
-    }
-    
-    @Test("DTXAPIClient URL validation edge cases")
-    func testURLValidationEdgeCases() async throws {
-        let client = DTXAPIClient()
-        
-        // Clean up any existing value
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        
-        // Test various edge cases for server URLs
-        let edgeCaseURLs = [
-            ("http://", "http://127.0.0.1:8001"),  // Invalid (no host), falls back to default
-            ("https://", "http://127.0.0.1:8001"),  // Invalid (no host), falls back to default
-            ("ftp://invalid-protocol.com", "http://127.0.0.1:8001"),  // Invalid protocol, falls back to default
-            ("   ", "http://127.0.0.1:8001"),  // whitespace only should fall back to default
-            ("http://valid.com:99999", "http://valid.com:99999"),  // very high port
-            ("http://127.0.0.1:0", "http://127.0.0.1:0")  // port 0
-        ]
-        
-        for (inputUrl, expectedUrl) in edgeCaseURLs {
-            client.setServerURL(inputUrl)
-            UserDefaults.standard.synchronize()
-            try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-            // Should not crash, and should handle whitespace appropriately
-            #expect(client.baseURL == expectedUrl)
-        }
-        
-        // Clean up after test
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-    }
-    
-    @Test("DTXAPIClient protocols conformance")
-    func testProtocolConformance() {
-        let client = DTXAPIClient()
-        
-        // Test that client conforms to expected protocols
-        #expect(client is DTXConfiguration)
-        #expect(client is DTXNetworking)
-        #expect(client is DTXFileOperations)
-        #expect(client is DTXDownloadOperations)
-    }
-    
-    @Test("DTXAPIClient UserDefaults integration")
-    func testUserDefaultsIntegration() async throws {
-        let client = DTXAPIClient()
-        let testKey = "DTXServerURL"
-        
-        // Clean up any existing value
-        UserDefaults.standard.removeObject(forKey: testKey)
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        
-        // Test setting and getting URL through UserDefaults
-        let testURL = "http://test-integration.com:9999"
-        client.setServerURL(testURL)
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        
-        let storedURL = UserDefaults.standard.string(forKey: testKey)
-        #expect(storedURL == testURL)
-        #expect(client.baseURL == testURL)
-        
-        // Test reset
-        client.resetToLocalServer()
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        let resetStoredURL = UserDefaults.standard.string(forKey: testKey)
-        #expect(resetStoredURL == nil)
-        #expect(client.baseURL == "http://127.0.0.1:8001")
-        
-        // Clean up
-        UserDefaults.standard.removeObject(forKey: testKey)
-        UserDefaults.standard.synchronize()
-    }
-    
-    @Test("DTXAPIClient handles concurrent configuration changes")
-    func testConcurrentConfigurationChanges() async throws {
-        let client = DTXAPIClient()
-        
-        // Clean up any existing value
-        UserDefaults.standard.removeObject(forKey: "DTXServerURL")
-        UserDefaults.standard.synchronize()
-        try await Task.sleep(nanoseconds: 5_000_000) // 0.005 seconds
-        
-        // Test concurrent URL changes
-        await withTaskGroup(of: Void.self) { group in
-            for i in 0..<10 {
-                group.addTask {
-                    client.setServerURL("http://server\(i).com")
-                    UserDefaults.standard.synchronize()
-                    _ = client.baseURL
-                    client.resetToLocalServer()
-                    UserDefaults.standard.synchronize()
-                }
-            }
-        }
-        
-        // Allow final operations to complete
-        try await Task.sleep(nanoseconds: 10_000_000) // 0.01 seconds
-        
-        // Should not crash and should have consistent final state
-        #expect(client.baseURL == "http://127.0.0.1:8001")
         
         // Clean up after test
         UserDefaults.standard.removeObject(forKey: "DTXServerURL")
